Add files via upload
This file containes the training data, testing data, and submission file. 
The train set contains gray-scale images of hand-drawn digits, from zero through nine. 
Each image is 28 pixels in height and 28 pixels in width, for a total of 784 pixels in total. This pixel-value is an integer between 0 and 255, inclusive.
The test set is the same as the training set, except that it does not contain the "label" column.